backend/requirements.txt
# FastAPI and server
fastapi==0.104.0
uvicorn==0.23.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.0.0

# Database
sqlalchemy==2.0.22
alembic==1.12.0
psycopg2-binary==2.9.9
asyncpg==0.28.0

# OpenAI
openai==1.1.1

# Image generation
pillow==10.0.1
python-dotenv==1.0.0

# Reddit API
praw==7.7.1

# Google Trends
pytrends==4.9.2

# Twitter API
tweepy==4.14.0

# Web scraping and data processing
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.1
numpy==1.26.0

# Testing
pytest==7.4.2
httpx==0.25.0

# Utilities
python-slugify==8.0.1
tenacity==8.2.3
schedule==1.2.0
aiohttp==3.8.6 